move ReadFileNoStat to its own file
This allows it to be built on all platforms
Fixes #221.
Also, updates the max size to 512k (instead of 500k) to match the comment.
